refactor(shared): 外移 shared_session_v2.rs 七个内联测试模块
- 全仓最大文件 7685→5349 行，纯搬运零行为变更
- 七个测试 mod 平铺为 #[path] 兄弟文件，mod 名与可见性不变
- 括号匹配定位真实 mod 边界（测试 mod 间夹有生产代码，不可按区间切）
- baseline 重生成：shared_session_v2 reduced -2336；新文件登记豁免
验证：cargo test --lib shared_session_v2 47 passed；cargo check 通过；gate 绿
