+++ src-tauri/src/shared_session_v2.rs
@@ -354,408 +354,6 @@ fn collaboration_mode_for_attempt(
     })
 }
 
-#[cfg(test)]
-mod execution_target_contract_tests {
-    use super::*;
-
-    #[test]
-    fn execution_target_input_accepts_canonical_local_and_rejects_catalog_disk() {
-        let local = serde_json::from_value::<ExecutionTargetInput>(json!({
-            "engine": "codex",
-            "providerProfileId": null,
-            "modelCatalogEntryId": "gpt-5.3-codex-spark",
-            "model": "gpt-5.3-codex-spark",
-            "reasoningEffort": null,
-            "providerProfileNameSnapshot": "本地配置",
-            "providerProfileSource": "local",
-            "runtimeCapabilityFingerprint": null
-        }))
-        .expect("canonical local target");
-        assert_eq!(
-            local.to_snapshot().provider_profile_source,
-            Some(CanonicalProviderProfileSource::Local)
-        );
-        assert_eq!(
-            local.model_catalog_entry_id.as_deref(),
-            Some("gpt-5.3-codex-spark")
-        );
-        assert_eq!(
-            local.to_snapshot().model_catalog_entry_id.as_deref(),
-            Some("gpt-5.3-codex-spark")
-        );
-
-        let error = serde_json::from_value::<ExecutionTargetInput>(json!({
-            "engine": "codex",
-            "providerProfileSource": "disk"
-        }))
-        .expect_err("catalog source must not cross canonical IPC boundary");
-        assert!(error.to_string().contains("unknown variant"));
-    }
-
-    #[test]
-    fn claude_session_identity_accepts_legacy_raw_and_canonical_prefix() {
-        assert_eq!(raw_claude_session_id("legacy-uuid"), Some("legacy-uuid"));
-        assert_eq!(
-            raw_claude_session_id("claude:canonical-uuid"),
-            Some("canonical-uuid")
-        );
-        assert_eq!(raw_claude_session_id("claude:"), None);
-    }
-
-    #[test]
-    fn execution_target_validation_rejects_mismatched_catalog_runtime_pair() {
-        // 从当前 generatedModelCatalog 动态取条目，避免模型目录漂移使用例失效。
-        let catalog =
-            crate::engine::status::get_local_engine_models_for_validation(EngineType::Codex)
-                .expect("codex local catalog");
-        let selected = catalog.first().expect("non-empty codex catalog");
-        let expected_runtime_model = if selected.model.trim().is_empty() {
-            selected.id.trim().to_string()
-        } else {
-            selected.model.trim().to_string()
-        };
-        let valid = ExecutionTargetInput {
-            engine: EngineType::Codex,
-            provider_profile_id: None,
-            model_catalog_entry_id: Some(selected.id.clone()),
-            model: Some(expected_runtime_model.clone()),
-            reasoning_effort: None,
-            provider_profile_name_snapshot: Some("本地配置".to_string()),
-            provider_profile_source: Some(CanonicalProviderProfileSource::Local),
-            runtime_capability_fingerprint: None,
-        };
-        assert_eq!(
-            validate_resolved_execution_target(&valid).expect("valid resolved local Codex pair"),
-            EngineType::Codex
-        );
-
-        let poisoned = ExecutionTargetInput {
-            model: Some("kimi-for-coding".to_string()),
-            ..valid
-        };
-        assert!(validate_resolved_execution_target(&poisoned)
-            .expect_err("mismatched runtime model must fail before the turn is persisted")
-            .contains(&format!(
-                "requires runtime model '{expected_runtime_model}'"
-            )));
-    }
-
-    #[test]
-    fn execution_target_validation_accepts_new_shared_cli_local_catalogs() {
-        for engine in [
-            EngineType::Kimi,
-            EngineType::Grok,
-            EngineType::OpenCode,
-            EngineType::Pi,
-        ] {
-            let catalog = crate::engine::status::get_local_engine_models_for_validation(engine)
-                .unwrap_or_else(|| panic!("missing local catalog for {engine:?}"));
-            let selected = catalog
-                .first()
-                .unwrap_or_else(|| panic!("empty local catalog for {engine:?}"));
-            let target = ExecutionTargetInput {
-                engine,
-                provider_profile_id: None,
-                model_catalog_entry_id: Some(selected.id.clone()),
-                model: Some(selected.model.clone()),
-                reasoning_effort: None,
-                provider_profile_name_snapshot: Some("本地配置".to_string()),
-                provider_profile_source: Some(CanonicalProviderProfileSource::Local),
-                runtime_capability_fingerprint: None,
-            };
-
-            assert_eq!(
-                validate_resolved_execution_target(&target)
-                    .unwrap_or_else(|error| panic!("{engine:?} local target rejected: {error}")),
-                engine
-            );
-        }
-    }
-
-    #[test]
-    fn resolved_execution_target_rejects_legacy_partial_identity() {
-        let partial = ExecutionTargetInput {
-            engine: EngineType::Codex,
-            provider_profile_id: None,
-            model_catalog_entry_id: None,
-            model: None,
-            reasoning_effort: None,
-            provider_profile_name_snapshot: None,
-            provider_profile_source: None,
-            runtime_capability_fingerprint: None,
-        };
-
-        assert!(validate_resolved_execution_target(&partial)
-            .expect_err("legacy partial target must fail closed")
-            .contains("providerProfileSource"),);
-    }
-
-    #[test]
-    fn resolved_execution_target_requires_source_to_match_provider_identity() {
-        let managed_with_local_source = ExecutionTargetInput {
-            engine: EngineType::Codex,
-            provider_profile_id: Some("provider-kimi".to_string()),
-            model_catalog_entry_id: Some("kimi-entry".to_string()),
-            model: Some("kimi-for-coding".to_string()),
-            reasoning_effort: None,
-            provider_profile_name_snapshot: Some("Kimi".to_string()),
-            provider_profile_source: Some(CanonicalProviderProfileSource::Local),
-            runtime_capability_fingerprint: None,
-        };
-
-        assert!(
-            validate_resolved_execution_target(&managed_with_local_source)
-                .expect_err("managed provider cannot claim local provenance")
-                .contains("must be 'managed'"),
-        );
-    }
-
-    #[test]
-    fn claude_shared_context_always_requires_exact_replay_echo() {
-        let target = ExecutionTargetInput {
-            engine: EngineType::Claude,
-            provider_profile_id: None,
-            model_catalog_entry_id: Some("claude-sonnet-4-5".to_string()),
-            model: Some("claude-sonnet-4-5".to_string()),
-            reasoning_effort: None,
-            provider_profile_name_snapshot: Some("本地配置".to_string()),
-            provider_profile_source: Some(CanonicalProviderProfileSource::Local),
-            runtime_capability_fingerprint: None,
-        };
-
-        assert!(context_capabilities(&target).strong_context_ack);
-    }
-
-    #[test]
-    fn codex_shared_context_uses_weak_portable_transcript() {
-        let target = ExecutionTargetInput {
-            engine: EngineType::Codex,
-            provider_profile_id: Some("compatible-provider".to_string()),
-            model_catalog_entry_id: Some("codex-model".to_string()),
-            model: Some("codex-model".to_string()),
-            reasoning_effort: None,
-            provider_profile_name_snapshot: Some("Compatible Provider".to_string()),
-            provider_profile_source: Some(CanonicalProviderProfileSource::Managed),
-            runtime_capability_fingerprint: Some("thread/inject_items".to_string()),
-        };
-
-        let capabilities = context_capabilities(&target);
-        assert!(capabilities.user_channel_transcript);
-        assert!(!capabilities.structured_history_import);
-        assert!(!capabilities.tool_history);
-        assert!(!capabilities.strong_context_ack);
-    }
-
-    #[test]
-    fn newly_supported_shared_engines_use_weak_user_channel_context() {
-        for engine in [
-            EngineType::Kimi,
-            EngineType::Grok,
-            EngineType::OpenCode,
-            EngineType::Pi,
-            EngineType::Qoder,
-        ] {
-            let target = ExecutionTargetInput {
-                engine,
-                provider_profile_id: None,
-                model_catalog_entry_id: Some("runtime-model".to_string()),
-                model: Some("runtime-model".to_string()),
-                reasoning_effort: None,
-                provider_profile_name_snapshot: Some("本地配置".to_string()),
-                provider_profile_source: Some(CanonicalProviderProfileSource::Local),
-                runtime_capability_fingerprint: None,
-            };
-            let capabilities = context_capabilities(&target);
-            assert!(capabilities.user_channel_transcript, "{engine:?}");
-            assert!(!capabilities.structured_history_import, "{engine:?}");
-            assert!(!capabilities.strong_context_ack, "{engine:?}");
-        }
-    }
-
-    #[test]
-    fn newly_supported_shared_engines_use_provider_scoped_runtime_keys() {
-        for (engine, local_suffix, managed_suffix) in [
-            (
-                EngineType::Kimi,
-                crate::engine::kimi_provider_profile::KIMI_LOCAL_PROVIDER_PROFILE_ID,
-                "provider-kimi",
-            ),
-            (
-                EngineType::Grok,
-                crate::engine::grok_provider_profile::GROK_LOCAL_PROVIDER_PROFILE_ID,
-                "provider-grok",
-            ),
-            (
-                EngineType::OpenCode,
-                crate::engine::opencode_provider_profile::OPENCODE_LOCAL_PROVIDER_PROFILE_ID,
-                "provider-opencode",
-            ),
-        ] {
-            assert_eq!(
-                provider_runtime_key_for_target("workspace-1", engine, None)
-                    .expect("local runtime key"),
-                format!("{}::workspace-1::{local_suffix}", engine.icon()),
-            );
-            assert_eq!(
-                provider_runtime_key_for_target("workspace-1", engine, Some(managed_suffix))
-                    .expect("managed runtime key"),
-                format!("{}::workspace-1::{managed_suffix}", engine.icon()),
-            );
-        }
-    }
-
-    #[test]
-    fn pi_shared_runtime_key_matches_native_ownership() {
-        assert_eq!(
-            provider_runtime_key_for_target("workspace-1", EngineType::Pi, None)
-                .expect("pi local runtime key"),
-            "workspace-1",
-        );
-        assert_eq!(
-            provider_runtime_key_for_target("workspace-1", EngineType::Pi, Some("custom"))
-                .expect("pi named runtime key"),
-            "workspace-1::pi::custom",
-        );
-    }
-
-    #[test]
-    fn qoder_shared_runtime_key_matches_native_ownership() {
-        // Qoder runtime key 必须携带 distribution；Global/CN 可在同一 workspace 并发。
-        assert_eq!(
-            provider_runtime_key_for_target("workspace-1", EngineType::Qoder, None)
-                .expect("qoder Global runtime key"),
-            "workspace-1::qoder::global",
-        );
-        let global_runtime_key = provider_runtime_key_for_target(
-            "workspace-1",
-            EngineType::Qoder,
-            Some(crate::engine::qoder_provider_profile::QODER_GLOBAL_PROVIDER_PROFILE_ID),
-        )
-        .expect("explicit qoder Global runtime key");
-        assert_eq!(global_runtime_key, "workspace-1::qoder::global");
-        let cn_runtime_key = provider_runtime_key_for_target(
-            "workspace-1",
-            EngineType::Qoder,
-            Some(crate::engine::qoder_provider_profile::QODER_CN_PROVIDER_PROFILE_ID),
-        )
-        .expect("qoder CN runtime key");
-        assert_eq!(cn_runtime_key, "workspace-1::qoder::cn",);
-        assert_ne!(global_runtime_key, cn_runtime_key);
-        assert!(provider_runtime_key_for_target(
-            "workspace-1",
-            EngineType::Qoder,
-            Some("unknown-qoder-profile"),
-        )
-        .is_err());
-    }
-
-    #[test]
-    fn qoder_runtime_only_catalog_accepts_legacy_and_explicit_distribution_targets() {
-        // Qoder 模型目录是 ACP runtime-only：发送路径禁止现场 probe，catalog 不可得
-        // 时按空目录 + Allow 放行（Session Switch Catalog Fetch Gate）。
-        let legacy_target = ExecutionTargetInput {
-            engine: EngineType::Qoder,
-            provider_profile_id: None,
-            model_catalog_entry_id: Some("qmodel_38max".to_string()),
-            model: Some("qmodel_38max".to_string()),
-            reasoning_effort: None,
-            provider_profile_name_snapshot: Some("本地配置".to_string()),
-            provider_profile_source: Some(CanonicalProviderProfileSource::Local),
-            runtime_capability_fingerprint: None,
-        };
-        assert_eq!(
-            validate_resolved_execution_target(&legacy_target)
-                .expect("qoder runtime-only catalog must not hard-fail"),
-            EngineType::Qoder
-        );
-
-        let legacy_sentinel = ExecutionTargetInput {
-            provider_profile_id: Some(
-                crate::engine::qoder_provider_profile::QODER_LOCAL_PROVIDER_PROFILE_ID.to_string(),
-            ),
-            provider_profile_name_snapshot: Some("Qoder Global".to_string()),
-            provider_profile_source: Some(CanonicalProviderProfileSource::Managed),
-            ..legacy_target.clone()
-        };
-        assert_eq!(
-            validate_resolved_execution_target(&legacy_sentinel)
-                .expect("legacy Qoder sentinel must remain Global-compatible"),
-            EngineType::Qoder
-        );
-
-        for (provider_profile_id, provider_name) in [
-            (
-                crate::engine::qoder_provider_profile::QODER_GLOBAL_PROVIDER_PROFILE_ID,
-                "Qoder Global",
-            ),
-            (
-                crate::engine::qoder_provider_profile::QODER_CN_PROVIDER_PROFILE_ID,
-                "Qoder CN",
-            ),
-        ] {
-            let target = ExecutionTargetInput {
-                provider_profile_id: Some(provider_profile_id.to_string()),
-                provider_profile_name_snapshot: Some(provider_name.to_string()),
-                provider_profile_source: Some(CanonicalProviderProfileSource::Managed),
-                ..legacy_target.clone()
-            };
-            assert_eq!(
-                validate_resolved_execution_target(&target)
-                    .unwrap_or_else(|error| panic!("{provider_name} target rejected: {error}")),
-                EngineType::Qoder
-            );
-        }
-
-        let unknown_target = ExecutionTargetInput {
-            provider_profile_id: Some("provider-qoder".to_string()),
-            provider_profile_name_snapshot: Some("Unknown Qoder".to_string()),
-            provider_profile_source: Some(CanonicalProviderProfileSource::Managed),
-            ..legacy_target
-        };
-        assert!(validate_resolved_execution_target(&unknown_target)
-            .expect_err("unknown Qoder profile must fail before Tx1")
-            .contains("QODER_DISTRIBUTION"));
-    }
-
-    #[test]
-    fn collaboration_mode_uses_attempt_model_and_clears_stale_reasoning() {
-        let target = ExecutionTargetInput {
-            engine: EngineType::Codex,
-            provider_profile_id: Some("minimax".to_string()),
-            model_catalog_entry_id: Some("minimax-m3".to_string()),
-            model: Some("MiniMax-M3".to_string()),
-            reasoning_effort: None,
-            provider_profile_name_snapshot: Some("MiniMax".to_string()),
-            provider_profile_source: Some(CanonicalProviderProfileSource::Managed),
-            runtime_capability_fingerprint: None,
-        };
-        let rewritten = collaboration_mode_for_attempt(
-            Some(json!({
-                "mode": "default",
-                "settings": {
-                    "model": "gpt-5.6-sol",
-                    "reasoning_effort": "high",
-                    "developer_instructions": "keep-me"
-                }
-            })),
-            &target,
-        )
-        .expect("collaboration mode");
-
-        assert_eq!(
-            rewritten.pointer("/settings/model").and_then(Value::as_str),
-            Some("MiniMax-M3")
-        );
-        assert!(rewritten.pointer("/settings/reasoning_effort").is_none());
-        assert_eq!(
-            rewritten
-                .pointer("/settings/developer_instructions")
-                .and_then(Value::as_str),
-            Some("keep-me")
-        );
-    }
-}
-
 /// commit_turn 的 outcome 输入。
 #[derive(Debug, Clone, Deserialize)]
 #[serde(rename_all = "camelCase")]
@@ -1240,62 +838,6 @@ fn resolve_dispatch_images(
         None
     } else {
         Some(from_refs)
-    }
-}
-
-#[cfg(test)]
-mod resolve_dispatch_images_tests {
-    use super::resolve_dispatch_images;
-    use crate::shared_event_log::canonical::types::{ArtifactRef, CanonicalUserInput};
-    use serde_json::Value;
-
-    fn artifact(locator: &str) -> ArtifactRef {
-        ArtifactRef {
-            artifact_id: "img-1".into(),
-            media_type: "image/png".into(),
-            size_bytes: Some(12),
-            sha256: "a".repeat(64),
-            locator: locator.into(),
-            redaction: None,
-            extra: Value::Object(Default::default()),
-        }
-    }
-
-    #[test]
-    fn prefers_explicit_param_over_image_refs() {
-        let input = CanonicalUserInput {
-            text: Some("hi".into()),
-            image_refs: Some(vec![artifact("/from/ref.png")]),
-            attachment_refs: None,
-            extra: Value::Object(Default::default()),
-        };
-        let resolved = resolve_dispatch_images(Some(vec!["/from/param.png".into()]), &input);
-        assert_eq!(resolved, Some(vec!["/from/param.png".into()]));
-    }
-
-    #[test]
-    fn falls_back_to_durable_image_refs_when_param_empty() {
-        let input = CanonicalUserInput {
-            text: Some("这是什么".into()),
-            image_refs: Some(vec![artifact("/Users/me/shot.png")]),
-            attachment_refs: None,
-            extra: Value::Object(Default::default()),
-        };
-        let resolved = resolve_dispatch_images(None, &input);
-        assert_eq!(resolved, Some(vec!["/Users/me/shot.png".into()]));
-        let resolved_empty = resolve_dispatch_images(Some(vec![]), &input);
-        assert_eq!(resolved_empty, Some(vec!["/Users/me/shot.png".into()]));
-    }
-
-    #[test]
-    fn returns_none_when_no_images_anywhere() {
-        let input = CanonicalUserInput {
-            text: Some("no images".into()),
-            image_refs: None,
-            attachment_refs: None,
-            extra: Value::Object(Default::default()),
-        };
-        assert_eq!(resolve_dispatch_images(None, &input), None);
     }
 }
 
@@ -2742,38 +2284,6 @@ fn validate_shared_session_workspace_owner(
     Ok(())
 }
 
-#[cfg(test)]
-mod shared_session_workspace_owner_tests {
-    use super::validate_shared_session_workspace_owner;
-
-    #[test]
-    fn workspace_owner_requires_exact_session_and_workspace_identity() {
-        assert!(validate_shared_session_workspace_owner(
-            "session-1",
-            "workspace-a",
-            "session-1",
-            "workspace-a",
-        )
-        .is_ok());
-        assert!(validate_shared_session_workspace_owner(
-            "session-1",
-            "workspace-a",
-            "session-1",
-            "workspace-b",
-        )
-        .expect_err("cross-workspace owner must fail closed")
-        .contains("shared-session-owner-mismatch"));
-        assert!(validate_shared_session_workspace_owner(
-            "session-2",
-            "workspace-a",
-            "session-1",
-            "workspace-a",
-        )
-        .expect_err("cross-session owner must fail closed")
-        .contains("shared-session-owner-mismatch"));
-    }
-}
-
 fn runtime_turn_id(response: &Value) -> Option<String> {
     response
         .pointer("/result/turn/id")
@@ -2900,369 +2410,6 @@ fn validate_runtime_dispatch_receipt(
         );
     }
     Ok(receipt.clone())
-}
-
-#[cfg(test)]
-mod runtime_dispatch_receipt_tests {
-    use super::*;
-    use crate::shared_event_log::{open, OpenOutcome};
-
-    fn durable_owner_for_receipt_test(
-        engine: EngineType,
-        provider_profile_id: Option<&str>,
-        model: &str,
-        reasoning_effort: Option<&str>,
-    ) -> DurableAttemptOwner {
-        let root = std::env::temp_dir().join(format!(
-            "mossx-shared-dispatch-receipt-{}",
-            uuid::Uuid::new_v4()
-        ));
-        std::fs::create_dir_all(&root).expect("create receipt test root");
-        let writer = match open(&root.join("shared-events.db")).expect("open receipt test store") {
-            OpenOutcome::Ready(writer) => writer,
-            OpenOutcome::ReadOnlyRecovery { reason, .. } => {
-                panic!("unexpected receipt test recovery store: {reason}")
-            }
-        };
-        let attempt_id = "attempt-receipt";
-        writer
-            .append_canonical_fact(
-                "receipt-session".to_string(),
-                CanonicalFact::TurnRequested(TurnRequestedFact {
-                    logical_turn_id: "logical-receipt".to_string(),
-                    attempt_id: attempt_id.to_string(),
-                    retry_of_attempt_id: None,
-                    input: CanonicalUserInput {
-                        text: Some("hello".to_string()),
-                        image_refs: None,
-                        attachment_refs: None,
-                        extra: Value::Object(Default::default()),
-                    },
-                    target: TurnExecutionSnapshot {
-                        engine: engine.icon().to_string(),
-                        provider_profile_id: provider_profile_id.map(str::to_string),
-                        model_catalog_entry_id: Some(model.to_string()),
-                        model: Some(model.to_string()),
-                        reasoning: reasoning_effort.map(|effort| ReasoningSelection {
-                            effort: effort.to_string(),
-                            extra: Value::Object(Default::default()),
-                        }),
-                        provider_profile_name_snapshot: Some(
-                            provider_profile_id.unwrap_or("本地配置").to_string(),
-                        ),
-                        provider_profile_source: Some(if provider_profile_id.is_some() {
-                            CanonicalProviderProfileSource::Managed
-                        } else {
-                            CanonicalProviderProfileSource::Local
-                        }),
-                        runtime_capability_fingerprint: None,
-                        extra: Value::Object(Default::default()),
-                    },
-                    requested_at: 1,
-                    extra: Value::Object(Default::default()),
-                }),
-            )
-            .expect("append receipt owner");
-        let owner =
-            durable_attempt_owner(&writer, "receipt-session", attempt_id).expect("durable owner");
-        writer.shutdown().expect("shutdown receipt test writer");
-        std::fs::remove_dir_all(root).expect("remove receipt test root");
-        owner
-    }
-
-    #[test]
-    fn managed_codex_receipt_requires_exact_provider_runtime_key() {
-        let owner = durable_owner_for_receipt_test(
-            EngineType::Codex,
-            Some("provider-kimi"),
-            "kimi-for-coding",
-            Some("high"),
-        );
-        let workspace_id = "workspace-managed";
-        let expected_runtime_key = crate::shared::codex_core::session_key_for_provider(
-            workspace_id,
-            Some("provider-kimi"),
-        );
-        let receipt = json!({
-            "mossxDispatchReceipt": {
-                "engine": "codex",
-                "providerProfileId": "provider-kimi",
-                "providerProfileSource": "managed",
-                "providerRuntimeKey": expected_runtime_key,
-                "model": "kimi-for-coding",
-                "reasoningEffort": "high",
-            }
-        });
-
-        assert!(validate_runtime_dispatch_receipt(&receipt, &owner, workspace_id).is_ok());
-
-        let mut poisoned = receipt;
-        poisoned["mossxDispatchReceipt"]["providerRuntimeKey"] =
-            Value::String("workspace-managed::different-provider".to_string());
-        assert!(
-            validate_runtime_dispatch_receipt(&poisoned, &owner, workspace_id)
-                .expect_err("wrong Runtime owner must fail closed")
-                .contains("Runtime key does not match")
-        );
-    }
-
-    #[test]
-    fn claude_receipt_accepts_local_and_managed_provider_identity() {
-        let local_owner =
-            durable_owner_for_receipt_test(EngineType::Claude, None, "claude-sonnet-4-5", None);
-        let local_receipt = json!({
-            "mossxDispatchReceipt": {
-                "engine": "claude",
-                "providerProfileId": null,
-                "providerProfileSource": "local",
-                "providerRuntimeKey": format!(
-                    "claude::workspace-local::{}",
-                    crate::engine::claude::CLAUDE_LOCAL_PROVIDER_PROFILE_ID,
-                ),
-                "model": "claude-sonnet-4-5",
-                "reasoningEffort": null,
-            }
-        });
-        assert!(
-            validate_runtime_dispatch_receipt(&local_receipt, &local_owner, "workspace-local",)
-                .is_ok()
-        );
-        let mut wrong_local_runtime = local_receipt;
-        wrong_local_runtime["mossxDispatchReceipt"]["providerRuntimeKey"] =
-            json!("claude::workspace-local::provider-anthropic");
-        assert!(validate_runtime_dispatch_receipt(
-            &wrong_local_runtime,
-            &local_owner,
-            "workspace-local",
-        )
-        .is_err());
-
-        let managed_owner = durable_owner_for_receipt_test(
-            EngineType::Claude,
-            Some("provider-anthropic"),
-            "claude-opus-4-1",
-            Some("high"),
-        );
-        assert!(validate_runtime_dispatch_receipt(
-            &json!({
-                "mossxDispatchReceipt": {
-                    "engine": "claude",
-                    "providerProfileId": "provider-anthropic",
-                    "providerProfileSource": "managed",
-                    "providerRuntimeKey": "claude::workspace-managed::provider-anthropic",
-                    "model": "claude-opus-4-1",
-                    "reasoningEffort": "high",
-                }
-            }),
-            &managed_owner,
-            "workspace-managed",
-        )
-        .is_ok());
-    }
-
-    #[test]
-    fn newly_supported_engine_receipts_accept_local_and_managed_identity() {
-        for (engine, model, managed_provider) in [
-            (EngineType::Kimi, "kimi-k2", "provider-kimi"),
-            (EngineType::Grok, "grok-code-fast-1", "provider-grok"),
-            (
-                EngineType::OpenCode,
-                "ccgui/opencode-model",
-                "provider-opencode",
-            ),
-        ] {
-            let local_owner = durable_owner_for_receipt_test(engine, None, model, None);
-            let local_runtime_key =
-                provider_runtime_key_for_target("workspace-local", engine, None)
-                    .expect("local runtime key");
-            assert!(validate_runtime_dispatch_receipt(
-                &json!({
-                    "mossxDispatchReceipt": {
-                        "engine": engine.icon(),
-                        "providerProfileId": null,
-                        "providerProfileSource": "local",
-                        "providerRuntimeKey": local_runtime_key,
-                        "model": model,
-                        "reasoningEffort": null,
-                    }
-                }),
-                &local_owner,
-                "workspace-local",
-            )
-            .is_ok());
-
-            let managed_owner =
-                durable_owner_for_receipt_test(engine, Some(managed_provider), model, Some("high"));
-            let managed_runtime_key = provider_runtime_key_for_target(
-                "workspace-managed",
-                engine,
-                Some(managed_provider),
-            )
-            .expect("managed runtime key");
-            assert!(validate_runtime_dispatch_receipt(
-                &json!({
-                    "mossxDispatchReceipt": {
-                        "engine": engine.icon(),
-                        "providerProfileId": managed_provider,
-                        "providerProfileSource": "managed",
-                        "providerRuntimeKey": managed_runtime_key,
-                        "model": model,
-                        "reasoningEffort": "high",
-                    }
-                }),
-                &managed_owner,
-                "workspace-managed",
-            )
-            .is_ok());
-        }
-    }
-
-    #[test]
-    fn qoder_receipts_preserve_legacy_global_and_isolate_distributions() {
-        let model = "qmodel_38max";
-        let workspace_id = "workspace-qoder";
-
-        let legacy_owner = durable_owner_for_receipt_test(EngineType::Qoder, None, model, None);
-        let legacy_runtime_key =
-            provider_runtime_key_for_target(workspace_id, EngineType::Qoder, None)
-                .expect("legacy Qoder Global runtime key");
-        assert!(validate_runtime_dispatch_receipt(
-            &json!({
-                "mossxDispatchReceipt": {
-                    "engine": "qoder",
-                    "providerProfileId": null,
-                    "providerProfileSource": "local",
-                    "providerRuntimeKey": legacy_runtime_key,
-                    "model": model,
-                    "reasoningEffort": null,
-                }
-            }),
-            &legacy_owner,
-            workspace_id,
-        )
-        .is_ok());
-
-        let global_profile_id =
-            crate::engine::qoder_provider_profile::QODER_GLOBAL_PROVIDER_PROFILE_ID;
-        let cn_profile_id = crate::engine::qoder_provider_profile::QODER_CN_PROVIDER_PROFILE_ID;
-        let global_owner = durable_owner_for_receipt_test(
-            EngineType::Qoder,
-            Some(global_profile_id),
-            model,
-            Some("high"),
-        );
-        let global_runtime_key = provider_runtime_key_for_target(
-            workspace_id,
-            EngineType::Qoder,
-            Some(global_profile_id),
-        )
-        .expect("Qoder Global runtime key");
-        let cn_runtime_key =
-            provider_runtime_key_for_target(workspace_id, EngineType::Qoder, Some(cn_profile_id))
-                .expect("Qoder CN runtime key");
-        assert_ne!(global_runtime_key, cn_runtime_key);
-
-        let global_receipt = json!({
-            "mossxDispatchReceipt": {
-                "engine": "qoder",
-                "providerProfileId": global_profile_id,
-                "providerProfileSource": "managed",
-                "providerRuntimeKey": global_runtime_key,
-                "model": model,
-                "reasoningEffort": "high",
-            }
-        });
-        assert!(
-            validate_runtime_dispatch_receipt(&global_receipt, &global_owner, workspace_id).is_ok()
-        );
-
-        let cn_owner = durable_owner_for_receipt_test(
-            EngineType::Qoder,
-            Some(cn_profile_id),
-            model,
-            Some("high"),
-        );
-        let cn_receipt = json!({
-            "mossxDispatchReceipt": {
-                "engine": "qoder",
-                "providerProfileId": cn_profile_id,
-                "providerProfileSource": "managed",
-                "providerRuntimeKey": cn_runtime_key,
-                "model": model,
-                "reasoningEffort": "high",
-            }
-        });
-        assert!(validate_runtime_dispatch_receipt(&cn_receipt, &cn_owner, workspace_id).is_ok());
-
-        let mut cross_distribution_receipt = global_receipt;
-        cross_distribution_receipt["mossxDispatchReceipt"]["providerRuntimeKey"] =
-            json!(cn_runtime_key);
-        assert!(validate_runtime_dispatch_receipt(
-            &cross_distribution_receipt,
-            &global_owner,
-            workspace_id,
-        )
-        .expect_err("Qoder CN runtime key must not satisfy a Global receipt")
-        .contains("Runtime key does not match"));
-    }
-
-    #[test]
-    fn dispatch_receipt_missing_or_mismatched_identity_fails_closed() {
-        let owner = durable_owner_for_receipt_test(
-            EngineType::Codex,
-            Some("provider-kimi"),
-            "kimi-for-coding",
-            Some("high"),
-        );
-        let workspace_id = "workspace-managed";
-        let expected_runtime_key = crate::shared::codex_core::session_key_for_provider(
-            workspace_id,
-            Some("provider-kimi"),
-        );
-        let valid_receipt = json!({
-            "mossxDispatchReceipt": {
-                "engine": "codex",
-                "providerProfileId": "provider-kimi",
-                "providerProfileSource": "managed",
-                "providerRuntimeKey": expected_runtime_key,
-                "model": "kimi-for-coding",
-                "reasoningEffort": "high",
-            }
-        });
-
-        assert!(
-            validate_runtime_dispatch_receipt(&json!({}), &owner, workspace_id)
-                .expect_err("missing receipt must fail closed")
-                .contains("receipt is missing")
-        );
-
-        for (field, poisoned_value) in [
-            ("engine", json!("claude")),
-            ("providerProfileId", json!("provider-other")),
-            ("providerProfileSource", json!("local")),
-            ("model", json!("gpt-5.3-codex-spark")),
-            ("reasoningEffort", json!("low")),
-            ("providerRuntimeKey", Value::Null),
-        ] {
-            let mut poisoned = valid_receipt.clone();
-            poisoned["mossxDispatchReceipt"][field] = poisoned_value;
-            assert!(
-                validate_runtime_dispatch_receipt(&poisoned, &owner, workspace_id).is_err(),
-                "{field} mismatch must fail closed"
-            );
-        }
-
-        let mut missing_model = valid_receipt;
-        missing_model["mossxDispatchReceipt"]
-            .as_object_mut()
-            .expect("receipt object")
-            .remove("model");
-        assert!(
-            validate_runtime_dispatch_receipt(&missing_model, &owner, workspace_id)
-                .expect_err("missing field must fail closed")
-                .contains("missing model")
-        );
-    }
 }
 
 fn typed_dispatch_error(code: &str, error: &str) -> String {
@@ -6178,1508 +5325,25 @@ pub(crate) async fn shared_session_v2_turn_state(
     }))
 }
 
-#[cfg(test)]
-mod legacy_import_tests {
-    use super::*;
-    use crate::shared_event_log::{open, Fidelity, OpenOutcome};
-
-    #[test]
-    fn legacy_snapshot_import_is_fingerprinted_and_idempotent() {
-        let root = std::env::temp_dir().join(format!(
-            "mossx-shared-legacy-import-{}-{}",
-            std::process::id(),
-            Uuid::new_v4()
-        ));
-        std::fs::create_dir_all(&root).expect("create test root");
-        let writer = match open(&root.join("events.db")).expect("open store") {
-            OpenOutcome::Ready(writer) => writer,
-            OpenOutcome::ReadOnlyRecovery { reason, .. } => {
-                panic!("unexpected recovery store: {reason}")
-            }
-        };
-        let source_path = root.join("log.jsonl");
-        let items = vec![
-            json!({
-                "id": "legacy-user-1",
-                "kind": "message",
-                "role": "user",
-                "text": "legacy question",
-                "turnId": "legacy-turn-1"
-            }),
-            json!({
-                "id": "legacy-assistant-1",
-                "kind": "message",
-                "role": "assistant",
-                "text": "legacy answer",
-                "turnId": "legacy-turn-1",
-                "isFinal": true
-            }),
-        ];
-
-        for _ in 0..2 {
-            import_legacy_snapshot_items(
-                &writer,
-                "legacy-session",
-                &source_path,
-                &items,
-                EngineType::Claude,
-                42,
-            )
-            .expect("import snapshot");
-        }
-
-        let events = writer
-            .events_for_session("legacy-session")
-            .expect("legacy events");
-        assert_eq!(events.len(), 2);
-        assert!(events
-            .iter()
-            .all(|event| event.fidelity == Fidelity::PresentationOnly));
-        let marker = writer
-            .legacy_import("legacy-session")
-            .expect("read marker")
-            .expect("marker");
-        assert_eq!(marker.status, "completed");
-        assert!(marker
-            .imported_through_marker
-            .as_deref()
-            .is_some_and(|value| value.starts_with("snapshot-items:2:sha256:")));
-
-        writer.shutdown().expect("shutdown writer");
-        std::fs::remove_dir_all(root).expect("remove test root");
-    }
-}
 
 #[cfg(test)]
-mod shared_interrupt_owner_tests {
-    use super::*;
-    use crate::shared_event_log::{open, OpenOutcome, SessionTargetUpdate};
-    use crate::shared_runtime_coordinator::{SharedRuntimeAttemptOwner, SharedRuntimeCoordinator};
-    use std::path::PathBuf;
-    use std::time::{SystemTime, UNIX_EPOCH};
-
-    fn open_test_writer(tag: &str) -> (PathBuf, SharedEventWriter) {
-        let nonce = SystemTime::now()
-            .duration_since(UNIX_EPOCH)
-            .map(|duration| duration.as_nanos())
-            .unwrap_or_default();
-        let root = std::env::temp_dir().join(format!(
-            "mossx-shared-interrupt-{tag}-{}-{nonce}",
-            std::process::id()
-        ));
-        std::fs::create_dir_all(&root).expect("create test root");
-        let writer = match open(&root.join("shared-events.db")).expect("open store") {
-            OpenOutcome::Ready(writer) => writer,
-            OpenOutcome::ReadOnlyRecovery { reason, .. } => {
-                panic!("unexpected recovery store: {reason}")
-            }
-        };
-        (root, writer)
-    }
-
-    fn target(engine: EngineType, provider: &str) -> ExecutionTargetInput {
-        ExecutionTargetInput {
-            engine,
-            provider_profile_id: Some(provider.to_string()),
-            model_catalog_entry_id: Some(format!("{provider}-catalog-model")),
-            model: Some(match engine {
-                EngineType::Claude => "claude-sonnet-4-5".to_string(),
-                EngineType::Codex => "gpt-5-codex".to_string(),
-                EngineType::Kimi => "kimi-k2".to_string(),
-                EngineType::Grok => "ccgui/grok-4.5".to_string(),
-                EngineType::OpenCode => "ccgui/opencode-model".to_string(),
-                EngineType::Pi => "auto".to_string(),
-                EngineType::Qoder => "qmodel_38max".to_string(),
-                EngineType::Gemini | EngineType::Dsh => "unsupported".to_string(),
-            }),
-            reasoning_effort: Some("medium".to_string()),
-            provider_profile_name_snapshot: Some(provider.to_string()),
-            provider_profile_source: Some(CanonicalProviderProfileSource::Managed),
-            runtime_capability_fingerprint: None,
-        }
-    }
-
-    fn assert_route(engine: EngineType, provider: &str) {
-        let session_id = format!("interrupt-{provider}");
-        let shared_thread_id = format!("shared:{session_id}");
-        let (root, writer) = open_test_writer(provider);
-        let begin = begin_turn_core(
-            &writer,
-            &session_id,
-            &target(engine, provider),
-            "hello".to_string(),
-            None,
-        )
-        .expect("begin");
-        let attempt_id = begin.attempt_id.expect("attempt");
-        let logical_turn_id = begin.logical_turn_id.expect("logical turn");
-        let binding_key = begin.binding_key;
-        let snapshot = begin.snapshot.expect("snapshot");
-        let binding_operation_id = durable_attempt_owner(&writer, &session_id, &attempt_id)
-            .expect("durable owner")
-            .binding_operation_id;
-        let coordinator = SharedRuntimeCoordinator::default();
-        coordinator
-            .register_attempt(SharedRuntimeAttemptOwner {
-                workspace_id: "ws-1".to_string(),
-                provider_runtime_key: provider_runtime_key_for_target(
-                    "ws-1",
-                    engine,
-                    Some(provider),
-                )
-                .expect("provider runtime key"),
-                shared_session_id: session_id,
-                shared_thread_id: shared_thread_id.clone(),
-                logical_turn_id,
-                attempt_id: attempt_id.clone(),
-                binding_key: binding_key.clone(),
-                binding_operation_id,
-                engine,
-                execution_target_snapshot: snapshot,
-                native_session_id: Some(format!("native-{provider}")),
-                runtime_turn_id: Some(format!("run-{provider}")),
-                context_marker: None,
-            })
-            .expect("register owner");
-
-        let route = resolve_shared_attempt_interrupt_route(
-            &writer,
-            &coordinator,
-            "ws-1",
-            &shared_thread_id,
-            &attempt_id,
-        )
-        .expect("resolve route");
-        assert_eq!(route.engine, engine);
-        assert_eq!(route.provider_profile_id.as_deref(), Some(provider));
-        assert_eq!(route.binding_key, binding_key);
-        // 与 SharedRuntimeCoordinator::normalize_native_session_identity 对齐：
-        // Qoder 额外带 distribution；其余 CLI 使用 engine: 前缀；Codex/Gemini/Dsh 保持 raw。
-        let expected_native_thread_id = match engine {
-            EngineType::Claude
-            | EngineType::Kimi
-            | EngineType::Pi
-            | EngineType::Grok
-            | EngineType::OpenCode => {
-                format!("{}:native-{provider}", engine.icon())
-            }
-            EngineType::Qoder => format!("qoder:{provider}:native-{provider}"),
-            EngineType::Codex | EngineType::Gemini | EngineType::Dsh => {
-                format!("native-{provider}")
-            }
-        };
-        assert_eq!(route.native_thread_id, expected_native_thread_id);
-        assert_eq!(route.runtime_turn_id, format!("run-{provider}"));
-
-        coordinator.remove_attempt(&attempt_id);
-        writer.shutdown().expect("shutdown writer");
-        std::fs::remove_dir_all(root).expect("remove test root");
-    }
-
-    #[test]
-    fn active_recovery_returns_exact_accepted_owner_envelope() {
-        let session_id = "recovery-active-owner";
-        let shared_thread_id = format!("shared:{session_id}");
-        let provider = "provider-active";
-        let active_target = target(EngineType::Codex, provider);
-        let (root, writer) = open_test_writer("recovery-active-owner");
-        let begin = begin_turn_core(
-            &writer,
-            session_id,
-            &active_target,
-            "hello".to_string(),
-            None,
-        )
-        .expect("begin");
-        let attempt_id = begin.attempt_id.expect("attempt");
-        let logical_turn_id = begin.logical_turn_id.expect("logical turn");
-        let binding_key = begin.binding_key;
-        let snapshot = begin.snapshot.expect("snapshot");
-        let binding_operation_id = durable_attempt_owner(&writer, session_id, &attempt_id)
-            .expect("durable owner")
-            .binding_operation_id;
-        accept_turn_core(
-            &writer,
-            session_id,
-            &attempt_id,
-            &logical_turn_id,
-            &active_target,
-            "native-active",
-        )
-        .expect("accept");
-        let coordinator = SharedRuntimeCoordinator::default();
-        coordinator
-            .register_attempt(SharedRuntimeAttemptOwner {
-                workspace_id: "ws-1".to_string(),
-                provider_runtime_key: provider_runtime_key_for_target(
-                    "ws-1",
-                    EngineType::Codex,
-                    Some(provider),
-                )
-                .expect("provider runtime key"),
-                shared_session_id: session_id.to_string(),
-                shared_thread_id: shared_thread_id.clone(),
-                logical_turn_id,
-                attempt_id: attempt_id.clone(),
-                binding_key: binding_key.clone(),
-                binding_operation_id,
-                engine: EngineType::Codex,
-                execution_target_snapshot: snapshot,
-                native_session_id: Some("native-active".to_string()),
-                runtime_turn_id: Some("run-active".to_string()),
-                context_marker: None,
-            })
-            .expect("register owner");
-        let evidence = unresolved_attempt_evidence(&writer, session_id, None)
-            .expect("read evidence")
-            .into_iter()
-            .find(|evidence| evidence.owner.requested.attempt_id == attempt_id)
-            .expect("unresolved accepted attempt");
-
-        assert_eq!(recovery_disposition(&evidence, &coordinator), "active");
-        let response =
-            active_recovery_response(&writer, &coordinator, "ws-1", &shared_thread_id, &evidence)
-                .expect("resolve active recovery")
-                .expect("active response");
-        assert_eq!(
-            response.get("attemptId").and_then(Value::as_str),
-            Some(attempt_id.as_str())
-        );
-        assert_eq!(
-            response.get("bindingKey").and_then(Value::as_str),
-            Some(binding_key.as_str())
-        );
-        assert_eq!(
-            response.get("nativeThreadId").and_then(Value::as_str),
-            Some("native-active")
-        );
-        assert_eq!(
-            response.get("runtimeTurnId").and_then(Value::as_str),
-            Some("run-active")
-        );
-        assert_eq!(
-            response
-                .pointer("/executionTargetSnapshot/providerProfileId")
-                .and_then(Value::as_str),
-            Some(provider)
-        );
-
-        coordinator.remove_attempt(&attempt_id);
-        writer.shutdown().expect("shutdown writer");
-        std::fs::remove_dir_all(root).expect("remove test root");
-    }
-
-    #[test]
-    fn preaccepted_runtime_owner_is_not_reported_active() {
-        let session_id = "recovery-preaccepted-owner";
-        let shared_thread_id = format!("shared:{session_id}");
-        let provider = "provider-preaccepted";
-        let (root, writer) = open_test_writer("recovery-preaccepted-owner");
-        let begin = begin_turn_core(
-            &writer,
-            session_id,
-            &target(EngineType::Codex, provider),
-            "hello".to_string(),
-            None,
-        )
-        .expect("begin");
-        let attempt_id = begin.attempt_id.expect("attempt");
-        let binding_operation_id = durable_attempt_owner(&writer, session_id, &attempt_id)
-            .expect("durable owner")
-            .binding_operation_id;
-        let coordinator = SharedRuntimeCoordinator::default();
-        coordinator
-            .register_attempt(SharedRuntimeAttemptOwner {
-                workspace_id: "ws-1".to_string(),
-                provider_runtime_key: provider_runtime_key_for_target(
-                    "ws-1",
-                    EngineType::Codex,
-                    Some(provider),
-                )
-                .expect("provider runtime key"),
-                shared_session_id: session_id.to_string(),
-                shared_thread_id: shared_thread_id.clone(),
-                logical_turn_id: begin.logical_turn_id.expect("logical turn"),
-                attempt_id: attempt_id.clone(),
-                binding_key: begin.binding_key,
-                binding_operation_id,
-                engine: EngineType::Codex,
-                execution_target_snapshot: begin.snapshot.expect("snapshot"),
-                native_session_id: Some("native-preaccepted".to_string()),
-                runtime_turn_id: Some("run-preaccepted".to_string()),
-                context_marker: None,
-            })
-            .expect("register owner");
-        let evidence = unresolved_attempt_evidence(&writer, session_id, None)
-            .expect("read evidence")
-            .into_iter()
-            .find(|evidence| evidence.owner.requested.attempt_id == attempt_id)
-            .expect("unresolved attempt");
-
-        assert_ne!(recovery_disposition(&evidence, &coordinator), "active");
-        assert!(active_recovery_response(
-            &writer,
-            &coordinator,
-            "ws-1",
-            &shared_thread_id,
-            &evidence,
-        )
-        .expect("resolve recovery")
-        .is_none());
-
-        coordinator.remove_attempt(&attempt_id);
-        writer.shutdown().expect("shutdown writer");
-        std::fs::remove_dir_all(root).expect("remove test root");
-    }
-
-    #[test]
-    fn shared_compaction_selected_target_comes_from_v2_store() {
-        let session_id = "compaction-v2-target";
-        let (root, writer) = open_test_writer("compaction-v2-target");
-        writer
-            .upsert_session_target(&SessionTargetUpdate {
-                session_id: session_id.to_string(),
-                schema_version: 2,
-                selected_target_json: json!({
-                    "engine": "claude",
-                    "providerProfileId": "provider-v2"
-                })
-                .to_string(),
-                updated_at: 1,
-            })
-            .expect("persist V2 Target");
-
-        let (engine, provider_profile_id) =
-            resolve_durable_shared_compaction_target(&writer, session_id)
-                .expect("resolve durable target");
-        assert_eq!(engine, EngineType::Claude);
-        assert_eq!(provider_profile_id.as_deref(), Some("provider-v2"));
-
-        writer.shutdown().expect("shutdown writer");
-        std::fs::remove_dir_all(root).expect("remove test root");
-    }
-
-    #[test]
-    fn shared_compaction_route_prefers_unresolved_attempt_owner() {
-        let session_id = "compaction-active-attempt";
-        let (root, writer) = open_test_writer("compaction-active-attempt");
-        let active_target = target(EngineType::Codex, "provider-active");
-        let begin = begin_turn_core(
-            &writer,
-            session_id,
-            &active_target,
-            "hello".to_string(),
-            None,
-        )
-        .expect("begin");
-        let attempt_id = begin.attempt_id.expect("attempt");
-        let logical_turn_id = begin.logical_turn_id.expect("logical turn");
-        accept_turn_core(
-            &writer,
-            session_id,
-            &attempt_id,
-            &logical_turn_id,
-            &active_target,
-            "native-active",
-        )
-        .expect("accept");
-
-        let route = resolve_shared_compaction_route_core(&writer, session_id, || {
-            panic!("selected Target must not override an unresolved Attempt owner")
-        })
-        .expect("resolve compaction route");
-        assert_eq!(route.engine, EngineType::Codex);
-        assert_eq!(
-            route.provider_profile_id.as_deref(),
-            Some("provider-active")
-        );
-        assert_eq!(route.native_thread_id, "native-active");
-        assert!(route.has_unresolved_attempt);
-
-        writer.shutdown().expect("shutdown writer");
-        std::fs::remove_dir_all(root).expect("remove test root");
-    }
-
-    #[test]
-    fn shared_compaction_route_uses_selected_target_after_commit() {
-        let session_id = "compaction-selected-target";
-        let (root, writer) = open_test_writer("compaction-selected-target");
-        let selected_target = target(EngineType::Codex, "provider-selected");
-        let begin = begin_turn_core(
-            &writer,
-            session_id,
-            &selected_target,
-            "hello".to_string(),
-            None,
-        )
-        .expect("begin");
-        let attempt_id = begin.attempt_id.expect("attempt");
-        let logical_turn_id = begin.logical_turn_id.expect("logical turn");
-        accept_turn_core(
-            &writer,
-            session_id,
-            &attempt_id,
-            &logical_turn_id,
-            &selected_target,
-            "native-selected",
-        )
-        .expect("accept");
-        commit_turn_core(
-            &writer,
-            session_id,
-            &attempt_id,
-            &logical_turn_id,
-            &selected_target,
-            None,
-            &CommitOutcomeInput {
-                status: "completed".to_string(),
-                error_code: None,
-                error_message: None,
-                stop_reason: None,
-            },
-            None,
-        )
-        .expect("commit");
-
-        let route = resolve_shared_compaction_route_core(&writer, session_id, || {
-            Ok((EngineType::Codex, Some("provider-selected".to_string())))
-        })
-        .expect("resolve selected route");
-        assert_eq!(route.engine, EngineType::Codex);
-        assert_eq!(route.native_thread_id, "native-selected");
-        assert!(!route.has_unresolved_attempt);
-
-        writer.shutdown().expect("shutdown writer");
-        std::fs::remove_dir_all(root).expect("remove test root");
-    }
-
-    #[test]
-    fn shared_compaction_route_preserves_selected_claude_provider() {
-        let session_id = "compaction-selected-claude";
-        let (root, writer) = open_test_writer("compaction-selected-claude");
-        let selected_target = target(EngineType::Claude, "provider-managed");
-        let begin = begin_turn_core(
-            &writer,
-            session_id,
-            &selected_target,
-            "hello".to_string(),
-            None,
-        )
-        .expect("begin");
-        let attempt_id = begin.attempt_id.expect("attempt");
-        let logical_turn_id = begin.logical_turn_id.expect("logical turn");
-        accept_turn_core(
-            &writer,
-            session_id,
-            &attempt_id,
-            &logical_turn_id,
-            &selected_target,
-            "claude:native-managed",
-        )
-        .expect("accept");
-        commit_turn_core(
-            &writer,
-            session_id,
-            &attempt_id,
-            &logical_turn_id,
-            &selected_target,
-            None,
-            &CommitOutcomeInput {
-                status: "completed".to_string(),
-                error_code: None,
-                error_message: None,
-                stop_reason: None,
-            },
-            None,
-        )
-        .expect("commit");
-
-        let route = resolve_shared_compaction_route_core(&writer, session_id, || {
-            Ok((EngineType::Claude, Some("provider-managed".to_string())))
-        })
-        .expect("resolve selected Claude route");
-        assert_eq!(route.engine, EngineType::Claude);
-        assert_eq!(
-            route.provider_profile_id.as_deref(),
-            Some("provider-managed")
-        );
-        assert_eq!(route.native_thread_id, "claude:native-managed");
-        assert!(!route.has_unresolved_attempt);
-
-        writer.shutdown().expect("shutdown writer");
-        std::fs::remove_dir_all(root).expect("remove test root");
-    }
-
-    #[test]
-    fn shared_compaction_route_rejects_unsupported_engine_before_runtime_lookup() {
-        let session_id = "compaction-unsupported";
-        let (root, writer) = open_test_writer("compaction-unsupported");
-
-        let error = resolve_shared_compaction_route_core(&writer, session_id, || {
-            Ok((EngineType::Kimi, Some("provider-kimi".to_string())))
-        })
-        .expect_err("unsupported engine must fail closed");
-        assert!(error.contains("shared-compaction-unsupported"));
-
-        writer.shutdown().expect("shutdown writer");
-        std::fs::remove_dir_all(root).expect("remove test root");
-    }
-
-    #[test]
-    fn shared_interrupt_route_isolates_same_engine_provider_owners() {
-        assert_route(EngineType::Claude, "provider-a");
-        assert_route(EngineType::Claude, "provider-b");
-        assert_route(EngineType::Codex, "provider-codex");
-        assert_route(EngineType::Kimi, "provider-kimi");
-        assert_route(EngineType::Grok, "provider-grok");
-        assert_route(EngineType::OpenCode, "provider-opencode");
-        assert_route(EngineType::Pi, "provider-pi");
-    }
-
-    #[test]
-    fn qoder_shared_interrupt_routes_isolate_global_and_cn_owners() {
-        let global_profile_id =
-            crate::engine::qoder_provider_profile::QODER_GLOBAL_PROVIDER_PROFILE_ID;
-        let cn_profile_id = crate::engine::qoder_provider_profile::QODER_CN_PROVIDER_PROFILE_ID;
-        assert_ne!(
-            shared_target_binding_key(EngineType::Qoder, Some(global_profile_id)),
-            shared_target_binding_key(EngineType::Qoder, Some(cn_profile_id)),
-        );
-        assert_ne!(
-            provider_runtime_key_for_target("ws-1", EngineType::Qoder, Some(global_profile_id))
-                .expect("Global runtime key"),
-            provider_runtime_key_for_target("ws-1", EngineType::Qoder, Some(cn_profile_id))
-                .expect("CN runtime key"),
-        );
-        assert_route(EngineType::Qoder, global_profile_id);
-        assert_route(EngineType::Qoder, cn_profile_id);
-    }
-
-    #[test]
-    fn qoder_unknown_distribution_is_rejected_before_turn_requested_is_written() {
-        let session_id = "qoder-unknown-distribution";
-        let (root, writer) = open_test_writer(session_id);
-
-        let outcome = begin_turn_core(
-            &writer,
-            session_id,
-            &target(EngineType::Qoder, "provider-qoder"),
-            "hello".to_string(),
-            None,
-        )
-        .expect("begin must return target-unavailable rather than write");
-
-        assert_eq!(outcome.status, BeginTurnStatus::TargetUnavailable);
-        assert!(outcome
-            .reason
-            .as_deref()
-            .is_some_and(|reason| reason.contains("QODER_DISTRIBUTION")));
-        assert!(outcome.binding_key.is_empty());
-        assert!(writer
-            .events_for_session(session_id)
-            .expect("read durable events")
-            .is_empty());
-        assert!(writer
-            .binding_state(
-                session_id,
-                &shared_target_binding_key(EngineType::Qoder, Some("provider-qoder")),
-            )
-            .expect("read durable binding")
-            .is_none());
-
-        writer.shutdown().expect("shutdown writer");
-        std::fs::remove_dir_all(root).expect("remove test root");
-    }
-
-    #[test]
-    fn committed_attempt_is_detected_without_a_live_runtime_owner() {
-        let session_id = "interrupt-already-committed";
-        let (root, writer) = open_test_writer("already-committed");
-        let selected_target = target(EngineType::Claude, "provider-committed");
-        let begin = begin_turn_core(
-            &writer,
-            session_id,
-            &selected_target,
-            "hello".to_string(),
-            None,
-        )
-        .expect("begin");
-        let attempt_id = begin.attempt_id.expect("attempt");
-        let logical_turn_id = begin.logical_turn_id.expect("logical turn");
-        commit_turn_core(
-            &writer,
-            session_id,
-            &attempt_id,
-            &logical_turn_id,
-            &selected_target,
-            None,
-            &CommitOutcomeInput {
-                status: "failed".to_string(),
-                error_code: Some("test-terminal".to_string()),
-                error_message: Some("terminal already committed".to_string()),
-                stop_reason: None,
-            },
-            None,
-        )
-        .expect("commit terminal");
-
-        assert!(committed_attempt_sequence(&writer, session_id, &attempt_id)
-            .expect("query commit")
-            .is_some());
-
-        writer.shutdown().expect("shutdown writer");
-        std::fs::remove_dir_all(root).expect("remove test root");
-    }
-
-    #[test]
-    fn durable_terminal_response_projects_committed_outcome_without_runtime_owner() {
-        let session_id = "await-terminal-committed";
-        let (root, writer) = open_test_writer("await-terminal-committed");
-        let selected_target = target(EngineType::Claude, "provider-committed");
-        let begin = begin_turn_core(
-            &writer,
-            session_id,
-            &selected_target,
-            "hello".to_string(),
-            None,
-        )
-        .expect("begin");
-        let attempt_id = begin.attempt_id.expect("attempt");
-        let logical_turn_id = begin.logical_turn_id.expect("logical turn");
-        let binding_key = begin.binding_key;
-        commit_turn_core(
-            &writer,
-            session_id,
-            &attempt_id,
-            &logical_turn_id,
-            &selected_target,
-            None,
-            &CommitOutcomeInput {
-                status: "failed".to_string(),
-                error_code: Some("test-terminal".to_string()),
-                error_message: Some("terminal already committed".to_string()),
-                stop_reason: None,
-            },
-            None,
-        )
-        .expect("commit terminal");
-
-        let response = committed_terminal_response(&writer, session_id, &attempt_id, &binding_key)
-            .expect("query durable terminal")
-            .expect("committed response");
-        assert_eq!(
-            response.get("status").and_then(Value::as_str),
-            Some("committed")
-        );
-        assert_eq!(
-            response
-                .pointer("/terminal/outcome")
-                .and_then(Value::as_str),
-            Some("failed")
-        );
-        assert_eq!(
-            response.get("bindingKey").and_then(Value::as_str),
-            Some(binding_key.as_str())
-        );
-
-        writer.shutdown().expect("shutdown writer");
-        std::fs::remove_dir_all(root).expect("remove test root");
-    }
-
-    #[test]
-    fn shared_interrupt_route_rejects_runtime_owner_target_drift() {
-        let session_id = "interrupt-drift";
-        let shared_thread_id = format!("shared:{session_id}");
-        let (root, writer) = open_test_writer("owner-drift");
-        let begin = begin_turn_core(
-            &writer,
-            session_id,
-            &target(EngineType::Codex, "provider-a"),
-            "hello".to_string(),
-            None,
-        )
-        .expect("begin");
-        let attempt_id = begin.attempt_id.expect("attempt");
-        let binding_operation_id = durable_attempt_owner(&writer, session_id, &attempt_id)
-            .expect("durable owner")
-            .binding_operation_id;
-        let mut poisoned_snapshot = begin.snapshot.expect("snapshot");
-        poisoned_snapshot.provider_profile_id = Some("provider-b".to_string());
-        let coordinator = SharedRuntimeCoordinator::default();
-        coordinator
-            .register_attempt(SharedRuntimeAttemptOwner {
-                workspace_id: "ws-1".to_string(),
-                provider_runtime_key: provider_runtime_key_for_target(
-                    "ws-1",
-                    EngineType::Codex,
-                    Some("provider-b"),
-                )
-                .expect("provider runtime key"),
-                shared_session_id: session_id.to_string(),
-                shared_thread_id: shared_thread_id.clone(),
-                logical_turn_id: begin.logical_turn_id.expect("logical turn"),
-                attempt_id: attempt_id.clone(),
-                binding_key: "codex:provider-b".to_string(),
-                binding_operation_id,
-                engine: EngineType::Codex,
-                execution_target_snapshot: poisoned_snapshot,
-                native_session_id: Some("native-b".to_string()),
-                runtime_turn_id: Some("run-b".to_string()),
-                context_marker: None,
-            })
-            .expect("register poisoned owner");
-
-        let error = resolve_shared_attempt_interrupt_route(
-            &writer,
-            &coordinator,
-            "ws-1",
-            &shared_thread_id,
-            &attempt_id,
-        )
-        .expect_err("owner drift must fail closed");
-        assert!(error.contains("shared-control-owner-mismatch"));
-
-        coordinator.remove_attempt(&attempt_id);
-        writer.shutdown().expect("shutdown writer");
-        std::fs::remove_dir_all(root).expect("remove test root");
-    }
-
-    #[test]
-    fn abandon_unresolved_attempt_commits_cancelled_and_clears_recovery() {
-        let session_id = "abandon-unresolved";
-        let provider = "provider-abandon";
-        let (root, writer) = open_test_writer("abandon-unresolved");
-        let begin = begin_turn_core(
-            &writer,
-            session_id,
-            &target(EngineType::Codex, provider),
-            "hello".to_string(),
-            None,
-        )
-        .expect("begin");
-        let attempt_id = begin.attempt_id.expect("attempt");
-        let binding_key = begin.binding_key.clone();
-        mark_recovery_core(
-            &writer,
-            session_id,
-            &binding_key,
-            EngineType::Codex,
-            Some(provider.to_string()),
-            Some("test-ambiguous"),
-        )
-        .expect("mark recovery");
-        assert_eq!(
-            provisioning_state_of(
-                &writer
-                    .binding_state(session_id, &binding_key)
-                    .expect("read binding")
-                    .expect("binding exists")
-            ),
-            PROVISIONING_RECOVERY_REQUIRED
-        );
-
-        let committed = abandon_unresolved_attempt_core(
-            &writer,
-            session_id,
-            &attempt_id,
-            "user-abandon-unresolved",
-        )
-        .expect("abandon");
-        assert_eq!(committed.binding_key, binding_key);
-        assert!(unresolved_attempt_evidence(&writer, session_id, None)
-            .expect("evidence")
-            .is_empty());
-        assert_ne!(
-            provisioning_state_of(
-                &writer
-                    .binding_state(session_id, &binding_key)
-                    .expect("read binding")
-                    .expect("binding exists")
-            ),
-            PROVISIONING_RECOVERY_REQUIRED
-        );
-
-        // begin 不应再被旧 recovery 挡住。
-        let next = begin_turn_core(
-            &writer,
-            session_id,
-            &target(EngineType::Codex, provider),
-            "again".to_string(),
-            None,
-        )
-        .expect("begin after abandon");
-        assert_eq!(next.status, BeginTurnStatus::Creating);
-
-        writer.shutdown().expect("shutdown writer");
-        std::fs::remove_dir_all(root).expect("remove test root");
-    }
-
-    #[test]
-    fn rebuild_binding_core_settles_single_unresolved() {
-        let session_id = "rebuild-single-unresolved";
-        let provider = "provider-multi";
-        let (root, writer) = open_test_writer("rebuild-single-unresolved");
-        let first = begin_turn_core(
-            &writer,
-            session_id,
-            &target(EngineType::Codex, provider),
-            "one".to_string(),
-            None,
-        )
-        .expect("begin first");
-        let binding_key = first.binding_key.clone();
-        let rebuilt = rebuild_binding_core(&writer, session_id, &binding_key).expect("rebuild");
-        assert_eq!(rebuilt.replaced_attempt_ids.len(), 1);
-        let binding = writer
-            .binding_state(session_id, &binding_key)
-            .expect("read")
-            .expect("binding");
-        assert_eq!(
-            read_native_context_trust(&binding),
-            NativeContextTrust::Dirty,
-            "rebuild must mark trust dirty"
-        );
-
-        writer.shutdown().expect("shutdown writer");
-        std::fs::remove_dir_all(root).expect("remove test root");
-    }
-
-    #[test]
-    fn rebuild_binding_core_allows_squad_worker_key() {
-        // Squad worker keys are first-class; rebuild must clear recovery-required
-        // without durable-key mismatch refusal.
-        let session_id = "rebuild-squad-worker-key";
-        let provider = "provider-squad";
-        let (root, writer) = open_test_writer("rebuild-squad-worker-key");
-        let squad_key = format!("squad:run-1:node-plan:claude:{provider}");
-        mark_recovery_core(
-            &writer,
-            session_id,
-            &squad_key,
-            EngineType::Claude,
-            Some(provider.to_string()),
-            Some("squad-worker-binding-recovery-required"),
-        )
-        .expect("mark squad recovery");
-        let rebuilt = rebuild_binding_core(&writer, session_id, &squad_key).expect("rebuild squad");
-        assert!(rebuilt.replaced_attempt_ids.is_empty());
-        let binding = writer
-            .binding_state(session_id, &squad_key)
-            .expect("read")
-            .expect("binding");
-        assert_eq!(
-            provisioning_state_of(&binding),
-            PROVISIONING_PREPARED,
-            "squad rebuild must return provisioning to prepared"
-        );
-
-        writer.shutdown().expect("shutdown writer");
-        std::fs::remove_dir_all(root).expect("remove test root");
-    }
-
-    #[test]
-    fn mark_recovery_and_failed_terminal_mark_native_context_trust_dirty() {
-        let session_id = "trust-dirty-on-failure";
-        let provider = "provider-trust";
-        let (root, writer) = open_test_writer("trust-dirty-on-failure");
-        let begin = begin_turn_core(
-            &writer,
-            session_id,
-            &target(EngineType::Claude, provider),
-            "原任务正文".to_string(),
-            None,
-        )
-        .expect("begin");
-        let attempt_id = begin.attempt_id.expect("attempt");
-        let binding_key = begin.binding_key.clone();
-
-        // 先标 trusted，模拟曾成功 accept。
-        set_native_context_trust(
-            &writer,
-            session_id,
-            &binding_key,
-            NativeContextTrust::Trusted,
-        )
-        .expect("set trusted");
-        assert_eq!(
-            read_native_context_trust(
-                &writer
-                    .binding_state(session_id, &binding_key)
-                    .expect("read")
-                    .expect("binding")
-            ),
-            NativeContextTrust::Trusted
-        );
-
-        mark_recovery_core(
-            &writer,
-            session_id,
-            &binding_key,
-            EngineType::Claude,
-            Some(provider.to_string()),
-            Some("runtime-delivery-ambiguous"),
-        )
-        .expect("mark recovery");
-        assert_eq!(
-            read_native_context_trust(
-                &writer
-                    .binding_state(session_id, &binding_key)
-                    .expect("read")
-                    .expect("binding")
-            ),
-            NativeContextTrust::Dirty
-        );
-
-        // 失败 terminal 也保持 dirty。
-        commit_runtime_snapshot_core(
-            &writer,
-            session_id,
-            &attempt_id,
-            RuntimeFinalSnapshot {
-                assistant_blocks: vec![],
-                assistant_text: None,
-                tool_calls: vec![],
-                tool_results: vec![],
-                artifacts: vec![],
-                provider_private_refs: vec![],
-                omissions: vec![],
-                outcome: OutcomeStatus::Failed,
-                error_code: Some("503".to_string()),
-                error_message: Some("No available accounts".to_string()),
-                stop_reason: None,
-            },
-            Some("claude:native-trust"),
-        )
-        .expect("failed terminal");
-        assert_eq!(
-            read_native_context_trust(
-                &writer
-                    .binding_state(session_id, &binding_key)
-                    .expect("read")
-                    .expect("binding")
-            ),
-            NativeContextTrust::Dirty
-        );
-
-        writer.shutdown().expect("shutdown writer");
-        std::fs::remove_dir_all(root).expect("remove test root");
-    }
-
-    #[test]
-    fn dirty_zero_transfer_needs_history_detects_original_task() {
-        // 编译层：destination-owned 清空 transfer 后，session_needs_history 仍为 true。
-        use crate::shared_context::{
-            is_zero_transfer_package, session_needs_history, CompileContextRequest,
-            RuntimeContextCapabilities,
-        };
-        use crate::shared_event_log::{Fidelity, StoredEvent};
-
-        let stored =
-            |sequence: i64, event_id: &str, fact_type: &str, attempt: &str, payload: Value| {
-                StoredEvent {
-                    session_id: "s-needs".to_string(),
-                    sequence,
-                    event_id: event_id.to_string(),
-                    fact_type: fact_type.to_string(),
-                    logical_turn_id: Some(format!("turn-{sequence}")),
-                    attempt_id: Some(attempt.to_string()),
-                    dedupe_key: None,
-                    payload_json: payload.to_string(),
-                    payload_checksum: format!("sha256:{event_id}"),
-                    fidelity: Fidelity::Canonical,
-                    committed_at: sequence,
-                }
-            };
-        let events = vec![
-            stored(
-                1,
-                "req-1",
-                "conversation.turnRequested",
-                "attempt-1",
-                json!({"input": {"text": "原任务正文请实现登录"}}),
-            ),
-            stored(
-                2,
-                "acc-1",
-                "conversation.turnAccepted",
-                "attempt-1",
-                json!({"bindingKey": "claude:provider-a"}),
-            ),
-        ];
-        let caps = RuntimeContextCapabilities {
-            native_delta: false,
-            structured_history_import: false,
-            native_clone: false,
-            user_channel_transcript: true,
-            tool_history: false,
-            image_history: false,
-            strong_context_ack: true,
-        };
-        let owned = compile_context(
-            &events,
-            &CompileContextRequest {
-                session_id: "s-needs".to_string(),
-                binding_key: "claude:provider-a".to_string(),
-                destination: json!({"engine": "claude"}),
-                destination_native_session_id: Some("claude:native-1".to_string()),
-                from_sequence_exclusive: None,
-                through_sequence_inclusive: None,
-                exclude_attempt_id: None,
-                capabilities: caps.clone(),
-                budget_estimated_tokens: None,
-            },
-        )
-        .expect("owned compile");
-        assert!(
-            is_zero_transfer_package(&owned),
-            "destination-owned should empty transfer"
-        );
-        let needs = session_needs_history(
-            &events,
-            &CompileContextRequest {
-                session_id: "s-needs".to_string(),
-                binding_key: "claude:provider-a".to_string(),
-                destination: json!({"engine": "claude"}),
-                destination_native_session_id: Some("claude:native-1".to_string()),
-                from_sequence_exclusive: None,
-                through_sequence_inclusive: None,
-                exclude_attempt_id: None,
-                capabilities: caps,
-                budget_estimated_tokens: None,
-            },
-        )
-        .expect("needs");
-        assert!(needs, "full rematerialize must see original task");
-        let rematerialized = compile_context(
-            &events,
-            &CompileContextRequest {
-                session_id: "s-needs".to_string(),
-                binding_key: "claude:provider-a".to_string(),
-                destination: json!({"engine": "claude"}),
-                destination_native_session_id: None,
-                from_sequence_exclusive: None,
-                through_sequence_inclusive: None,
-                exclude_attempt_id: None,
-                capabilities: RuntimeContextCapabilities {
-                    native_delta: false,
-                    structured_history_import: false,
-                    native_clone: false,
-                    user_channel_transcript: true,
-                    tool_history: false,
-                    image_history: false,
-                    strong_context_ack: true,
-                },
-                budget_estimated_tokens: None,
-            },
-        )
-        .expect("rematerialize");
-        assert!(!is_zero_transfer_package(&rematerialized));
-        assert!(rematerialized
-            .prompt_prefix
-            .contains("原任务正文请实现登录"));
-    }
-
-    #[test]
-    fn dirty_non_zero_continue_only_package_still_needs_full_rematerialize() {
-        // 图1 P0：accepted 之后有未 Accepted 的「继续」turnRequested → 增量 package 非空
-        // 但缺原任务；dirty 时仍必须判定 needs_history 并全量 rematerialize。
-        use crate::shared_context::{
-            is_zero_transfer_package, session_needs_history, CompileContextRequest,
-            RuntimeContextCapabilities,
-        };
-        use crate::shared_event_log::{Fidelity, StoredEvent};
-
-        let stored =
-            |sequence: i64, event_id: &str, fact_type: &str, attempt: &str, payload: Value| {
-                StoredEvent {
-                    session_id: "s-continue-only".to_string(),
-                    sequence,
-                    event_id: event_id.to_string(),
-                    fact_type: fact_type.to_string(),
-                    logical_turn_id: Some(format!("turn-{sequence}")),
-                    attempt_id: Some(attempt.to_string()),
-                    dedupe_key: None,
-                    payload_json: payload.to_string(),
-                    payload_checksum: format!("sha256:{event_id}"),
-                    fidelity: Fidelity::Canonical,
-                    committed_at: sequence,
-                }
-            };
-        let events = vec![
-            stored(
-                1,
-                "req-orig",
-                "conversation.turnRequested",
-                "attempt-orig",
-                json!({"input": {"text": "原任务：实现登录并写测试"}}),
-            ),
-            stored(
-                2,
-                "acc-orig",
-                "conversation.turnAccepted",
-                "attempt-orig",
-                json!({"bindingKey": "claude:provider-a"}),
-            ),
-            // 失败轮：只有 turnRequested「继续」，无 turnAccepted → 不 destination-owned
-            stored(
-                3,
-                "req-c1",
-                "conversation.turnRequested",
-                "attempt-c1",
-                json!({"input": {"text": "继续"}}),
-            ),
-            stored(
-                4,
-                "req-c2",
-                "conversation.turnRequested",
-                "attempt-c2",
-                json!({"input": {"text": "继续"}}),
-            ),
-        ];
-        let caps = RuntimeContextCapabilities {
-            native_delta: false,
-            structured_history_import: false,
-            native_clone: false,
-            user_channel_transcript: true,
-            tool_history: false,
-            image_history: false,
-            strong_context_ack: true,
-        };
-        // 模拟 accepted_through=2 后的增量 compile（当前 attempt-c3 排除）
-        let incremental = compile_context(
-            &events,
-            &CompileContextRequest {
-                session_id: "s-continue-only".to_string(),
-                binding_key: "claude:provider-a".to_string(),
-                destination: json!({"engine": "claude"}),
-                destination_native_session_id: Some("claude:native-1".to_string()),
-                from_sequence_exclusive: Some(2),
-                through_sequence_inclusive: Some(4),
-                exclude_attempt_id: Some("attempt-c3".to_string()),
-                capabilities: caps.clone(),
-                budget_estimated_tokens: None,
-            },
-        )
-        .expect("incremental");
-        assert!(
-            !is_zero_transfer_package(&incremental),
-            "continue-only package is non-empty"
-        );
-        assert!(
-            !incremental.prompt_prefix.contains("原任务"),
-            "incremental must NOT contain original task"
-        );
-        assert!(
-            incremental.prompt_prefix.contains("继续"),
-            "incremental only has continues"
-        );
-        let needs = session_needs_history(
-            &events,
-            &CompileContextRequest {
-                session_id: "s-continue-only".to_string(),
-                binding_key: "claude:provider-a".to_string(),
-                destination: json!({"engine": "claude"}),
-                destination_native_session_id: Some("claude:native-1".to_string()),
-                from_sequence_exclusive: Some(2),
-                through_sequence_inclusive: Some(4),
-                exclude_attempt_id: Some("attempt-c3".to_string()),
-                capabilities: caps.clone(),
-                budget_estimated_tokens: None,
-            },
-        )
-        .expect("needs");
-        assert!(needs, "full history still needed");
-        let full = compile_context(
-            &events,
-            &CompileContextRequest {
-                session_id: "s-continue-only".to_string(),
-                binding_key: "claude:provider-a".to_string(),
-                destination: json!({"engine": "claude"}),
-                destination_native_session_id: None,
-                from_sequence_exclusive: None,
-                through_sequence_inclusive: Some(4),
-                exclude_attempt_id: Some("attempt-c3".to_string()),
-                capabilities: caps,
-                budget_estimated_tokens: None,
-            },
-        )
-        .expect("full rematerialize");
-        assert!(full.prompt_prefix.contains("原任务：实现登录并写测试"));
-        assert!(full.prompt_prefix.contains("继续"));
-    }
-
-    #[test]
-    fn missing_native_context_trust_field_defaults_dirty() {
-        let (_, writer) = open_test_writer("legacy-trust-default");
-        let session_id = "legacy-trust-session";
-        let binding_key = "claude:legacy";
-        writer
-            .upsert_binding_state(&BindingStateUpdate {
-                session_id: session_id.to_string(),
-                binding_key: binding_key.to_string(),
-                engine: "claude".to_string(),
-                provider_profile_id: Some("legacy".to_string()),
-                native_session_id: Some("claude:native-legacy".to_string()),
-                accepted_through_sequence: Some(3),
-                committed_through_sequence: Some(3),
-                // 无 nativeContextTrust 字段
-                provisioning_json: Some(json!({"state": "ready", "updatedAt": 1}).to_string()),
-                pending_delivery_json: None,
-                availability: "ready".to_string(),
-                updated_at: 1,
-            })
-            .expect("upsert");
-        let row = writer
-            .binding_state(session_id, binding_key)
-            .expect("read")
-            .expect("row");
-        assert_eq!(
-            read_native_context_trust(&row),
-            NativeContextTrust::Dirty,
-            "legacy missing field must fail-closed to dirty"
-        );
-        writer.shutdown().expect("shutdown");
-    }
-}
-
+#[path = "shared_session_v2_execution_target_contract_tests.rs"]
+mod execution_target_contract_tests;
 #[cfg(test)]
-mod native_continuation_import_tests {
-    use super::*;
-    use crate::native_history::{
-        ContextSourceEntry, NativeHistoryEngine, NativeHistoryFidelity, NativeHistoryReadResult,
-        NativeHistorySource,
-    };
-    use crate::shared_context::{compile_native_context, CompileNativeContextRequest};
-
-    #[test]
-    fn codex_projection_preserves_raw_tool_items_and_counts_unimportable_entries() {
-        let source = NativeHistorySource {
-            session_id: "codex:source".to_string(),
-            native_session_id: "source".to_string(),
-            engine: NativeHistoryEngine::Codex,
-            provider_profile_id: Some("provider-a".to_string()),
-        };
-        let package = compile_native_context(&CompileNativeContextRequest {
-            session_id: source.session_id.clone(),
-            binding_key: "continuation:op".to_string(),
-            destination: json!({"engine": "codex"}),
-            source,
-            history: NativeHistoryReadResult {
-                reader_id: "codex-rollout-jsonl/v1".to_string(),
-                source_fingerprint: "sha256:source".to_string(),
-                through_cursor: "jsonl-v1:1:sha256:source".to_string(),
-                entries: vec![
-                    ContextSourceEntry {
-                        source_entry_id: "tool".to_string(),
-                        occurred_at: None,
-                        role: "tool".to_string(),
-                        blocks: vec![json!({
-                            "kind": "native-block",
-                            "value": {
-                                "type": "function_call",
-                                "name": "shell",
-                                "arguments": "{}",
-                                "call_id": "call-1"
-                            }
-                        })],
-                        provenance: json!({}),
-                        fidelity: NativeHistoryFidelity::Semantic,
-                    },
-                    ContextSourceEntry {
-                        source_entry_id: "control".to_string(),
-                        occurred_at: None,
-                        role: "control".to_string(),
-                        blocks: vec![json!({"kind": "native-block", "value": {"type": "unknown"}})],
-                        provenance: json!({}),
-                        fidelity: NativeHistoryFidelity::Lossy,
-                    },
-                ],
-                fidelity: NativeHistoryFidelity::Semantic,
-                omissions: Vec::new(),
-            },
-            capabilities: RuntimeContextCapabilities {
-                native_delta: false,
-                structured_history_import: true,
-                native_clone: false,
-                user_channel_transcript: true,
-                tool_history: true,
-                image_history: false,
-                strong_context_ack: true,
-            },
-            budget_estimated_tokens: None,
-        })
-        .expect("compile");
-
-        let (items, dropped) = codex_import_projection(&package);
-        assert_eq!(items[0]["role"], "user");
-        assert!(items[0]["content"][0]["text"]
-            .as_str()
-            .is_some_and(|text| text.starts_with("MOSSX_CONTEXT_PACKAGE:")));
-        assert_eq!(items[1]["type"], "function_call");
-        let package_marker = items[0]["content"][0]["text"]
-            .as_str()
-            .expect("package marker");
-        let accepted_marker = items
-            .last()
-            .and_then(|item| item["content"][0]["text"].as_str())
-            .expect("accepted marker");
-        assert_eq!(
-            accepted_marker,
-            package_marker.replacen("MOSSX_CONTEXT_PACKAGE:", "MOSSX_CONTEXT_ACCEPTED:", 1)
-        );
-        assert_eq!(dropped, 1);
-        assert!(
-            items
-                .iter()
-                .all(|item| item.get("role").and_then(Value::as_str) != Some("control")),
-            "control roles must not be injected as messages"
-        );
-    }
-
-    #[test]
-    fn codex_import_projection_drops_control_role_text_messages() {
-        // DeepSeek 等兼容 API：unknown variant `control`，只认 user/assistant/system/developer。
-        let source = NativeHistorySource {
-            session_id: "codex:source".to_string(),
-            native_session_id: "source".to_string(),
-            engine: NativeHistoryEngine::Codex,
-            provider_profile_id: Some("provider-a".to_string()),
-        };
-        let package = compile_native_context(&CompileNativeContextRequest {
-            session_id: source.session_id.clone(),
-            binding_key: "continuation:op".to_string(),
-            destination: json!({"engine": "codex"}),
-            source,
-            history: NativeHistoryReadResult {
-                reader_id: "codex-rollout-jsonl/v1".to_string(),
-                source_fingerprint: "sha256:source".to_string(),
-                through_cursor: "jsonl-v1:2:sha256:source".to_string(),
-                entries: vec![
-                    ContextSourceEntry {
-                        source_entry_id: "u1".to_string(),
-                        occurred_at: None,
-                        role: "user".to_string(),
-                        blocks: vec![json!({"kind": "text", "text": "hello from user"})],
-                        provenance: json!({}),
-                        fidelity: NativeHistoryFidelity::Semantic,
-                    },
-                    ContextSourceEntry {
-                        source_entry_id: "control-meta".to_string(),
-                        occurred_at: None,
-                        role: "control".to_string(),
-                        blocks: vec![json!({
-                            "kind": "text",
-                            "text": "session meta that must not become a control message"
-                        })],
-                        provenance: json!({}),
-                        fidelity: NativeHistoryFidelity::Lossy,
-                    },
-                    ContextSourceEntry {
-                        source_entry_id: "a1".to_string(),
-                        occurred_at: None,
-                        role: "assistant".to_string(),
-                        blocks: vec![json!({"kind": "text", "text": "hello from assistant"})],
-                        provenance: json!({}),
-                        fidelity: NativeHistoryFidelity::Semantic,
-                    },
-                ],
-                fidelity: NativeHistoryFidelity::Semantic,
-                omissions: Vec::new(),
-            },
-            capabilities: RuntimeContextCapabilities {
-                native_delta: false,
-                structured_history_import: true,
-                native_clone: false,
-                user_channel_transcript: true,
-                tool_history: true,
-                image_history: false,
-                strong_context_ack: true,
-            },
-            budget_estimated_tokens: None,
-        })
-        .expect("compile");
-
-        let (items, dropped) = codex_import_projection(&package);
-        let roles: Vec<&str> = items
-            .iter()
-            .filter_map(|item| item.get("role").and_then(Value::as_str))
-            .collect();
-        assert!(roles.contains(&"user"));
-        assert!(roles.contains(&"assistant"));
-        assert!(
-            !roles.iter().any(|role| *role == "control"),
-            "control text must be dropped, got roles={roles:?}"
-        );
-        assert!(
-            dropped >= 1,
-            "control entry should count as dropped, dropped={dropped}"
-        );
-        assert!(
-            items.iter().any(|item| {
-                item.get("role") == Some(&json!("user"))
-                    && item["content"][0]["text"]
-                        .as_str()
-                        .is_some_and(|text| text.contains("hello from user"))
-            }),
-            "user text preserved"
-        );
-        assert!(
-            items.iter().any(|item| {
-                item.get("role") == Some(&json!("assistant"))
-                    && item["content"][0]["text"]
-                        .as_str()
-                        .is_some_and(|text| text.contains("hello from assistant"))
-            }),
-            "assistant text preserved"
-        );
-    }
-
-    #[test]
-    fn codex_zero_delta_projection_does_not_create_marker_only_import() {
-        let source = NativeHistorySource {
-            session_id: "codex:source".to_string(),
-            native_session_id: "source".to_string(),
-            engine: NativeHistoryEngine::Codex,
-            provider_profile_id: Some("provider-a".to_string()),
-        };
-        // 编译器对空 history fail-closed（d528fc91c），因此用一条有效 entry 编译、
-        // 再清空 delta 来构造 zero-transfer 包，守住「空 delta 不产 marker-only 导入」。
-        let mut package = compile_native_context(&CompileNativeContextRequest {
-            session_id: source.session_id.clone(),
-            binding_key: "continuation:op".to_string(),
-            destination: json!({"engine": "codex"}),
-            source,
-            history: NativeHistoryReadResult {
-                reader_id: "codex-rollout-jsonl/v1".to_string(),
-                source_fingerprint: "sha256:source".to_string(),
-                through_cursor: "jsonl-v1:1:sha256:source".to_string(),
-                entries: vec![ContextSourceEntry {
-                    source_entry_id: "user-1".to_string(),
-                    occurred_at: None,
-                    role: "user".to_string(),
-                    blocks: vec![json!({"text": "hello"})],
-                    provenance: json!({}),
-                    fidelity: NativeHistoryFidelity::Semantic,
-                }],
-                fidelity: NativeHistoryFidelity::Semantic,
-                omissions: Vec::new(),
-            },
-            capabilities: RuntimeContextCapabilities {
-                native_delta: false,
-                structured_history_import: true,
-                native_clone: false,
-                user_channel_transcript: true,
-                tool_history: true,
-                image_history: false,
-                strong_context_ack: true,
-            },
-            budget_estimated_tokens: None,
-        })
-        .expect("compile projection package");
-        package.delta.clear();
-
-        let (items, dropped) = codex_import_projection(&package);
-        assert!(items.is_empty());
-        assert_eq!(dropped, 0);
-    }
-}
+#[path = "shared_session_v2_resolve_dispatch_images_tests.rs"]
+mod resolve_dispatch_images_tests;
+#[cfg(test)]
+#[path = "shared_session_v2_shared_session_workspace_owner_tests.rs"]
+mod shared_session_workspace_owner_tests;
+#[cfg(test)]
+#[path = "shared_session_v2_runtime_dispatch_receipt_tests.rs"]
+mod runtime_dispatch_receipt_tests;
+#[cfg(test)]
+#[path = "shared_session_v2_legacy_import_tests.rs"]
+mod legacy_import_tests;
+#[cfg(test)]
+#[path = "shared_session_v2_shared_interrupt_owner_tests.rs"]
+mod shared_interrupt_owner_tests;
+#[cfg(test)]
+#[path = "shared_session_v2_native_continuation_import_tests.rs"]
+mod native_continuation_import_tests;
